requirements.txt
crccheck >= 0.6
pyserial
pyvesc_fix
PyCRC
serial
